Also move the DefaultTypeProvider to m-resolver-provider to remove duplication
