+++ maven-resolver-provider/src/main/java/org/apache/maven/repository/internal/type/DefaultTypeProvider.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.maven.internal.impl;
+package org.apache.maven.repository.internal.type;
 
 import javax.inject.Named;
 
@@ -26,25 +26,29 @@ import java.util.Collection;
 import org.apache.maven.api.Language;
 import org.apache.maven.api.Type;
 import org.apache.maven.api.spi.TypeProvider;
-import org.apache.maven.repository.internal.type.DefaultType;
 
 @Named
 public class DefaultTypeProvider implements TypeProvider {
+    @SuppressWarnings({"rawtypes", "unchecked"})
     @Override
     public Collection<Type> provides() {
+        return (Collection) types();
+    }
+
+    public Collection<DefaultType> types() {
         return Arrays.asList(
                 new DefaultType("bom", Language.NONE, "pom", null, false, false),
-                new DefaultType("ear", Language.JAVA_FAMILY, "ear", null, false, true),
-                new DefaultType("ejb-client", Language.JAVA_FAMILY, "jar", "client", true, false),
-                new DefaultType("ejb", Language.JAVA_FAMILY, "jar", null, true, false),
+                new DefaultType("pom", Language.NONE, "pom", null, false, false),
+                new DefaultType("maven-plugin", Language.JAVA_FAMILY, "jar", null, true, false),
                 new DefaultType("jar", Language.JAVA_FAMILY, "jar", null, true, false),
+                new DefaultType("ejb", Language.JAVA_FAMILY, "jar", null, true, false),
+                new DefaultType("ejb-client", Language.JAVA_FAMILY, "jar", "client", true, false),
+                new DefaultType("test-jar", Language.JAVA_FAMILY, "jar", "tests", true, false),
                 new DefaultType("javadoc", Language.JAVA_FAMILY, "jar", "javadoc", true, false),
                 new DefaultType("java-source", Language.JAVA_FAMILY, "jar", "sources", false, false),
-                new DefaultType("maven-plugin", Language.JAVA_FAMILY, "jar", null, true, false),
-                new DefaultType("par", Language.JAVA_FAMILY, "par", null, false, true),
-                new DefaultType("pom", Language.NONE, "pom", null, false, false),
+                new DefaultType("war", Language.JAVA_FAMILY, "war", null, false, true),
+                new DefaultType("ear", Language.JAVA_FAMILY, "ear", null, false, true),
                 new DefaultType("rar", Language.JAVA_FAMILY, "rar", null, false, true),
-                new DefaultType("test-jar", Language.JAVA_FAMILY, "jar", "tests", true, false),
-                new DefaultType("war", Language.JAVA_FAMILY, "war", null, false, true));
+                new DefaultType("par", Language.JAVA_FAMILY, "par", null, false, true));
     }
 }
